Backport 1.12.2
fix


fix2


fix


fix2


fix config, fix optifine


fix missing lib


fix


make simpler


fix mixin json


fix


fix2


fix3


fix


fix2


hopefully fix spaces in file name


is this file even used on 1.12?


fix playSound on old mc


fix


fix


fix player nametags


fix getEntities on -=1.12.2


fix missing function

+++ common/src/main/java/xyz/wagyourtail/jsmacros/client/mixins/access/MixinHandledScreen.java
@@ -40,7 +40,7 @@ public class MixinHandledScreen<T extends ScreenHandler> extends Screen implemen
     protected int y;
 
     @Override
-    public Slot jsmacros_getSlotUnder(double x, double y) {
+    public Slot jsmacros_getSlotUnder(int x, int y) {
         return getSlotAt(x, y);
     }
